+++ src/Manager/HasLoggers.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Nip\Logger\Manager;
 
 use Nip\Logger\Logger;
@@ -7,16 +9,14 @@ use Psr\Log\LoggerInterface;
 
 /**
  * Trait HasLoggers
+ *
  * @package Nip\Logger\Manager
  */
 trait HasLoggers
 {
-    /**
-     * @param $logger
-     * @return Logger|LoggerInterface
-     */
-    protected function createLogger($logger)
+    protected function createLogger(mixed $logger): Logger|LoggerInterface
     {
         return new Logger($logger);
     }
 }
+