+++ src/Manager/MonologWrappers.php
@@ -1,8 +1,9 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Nip\Logger\Manager;
 
-use League\Container\Exception\NotFoundException;
 use Monolog\Formatter\LineFormatter;
 use Monolog\Handler\FormattableHandlerInterface;
 use Monolog\Handler\HandlerInterface;
@@ -12,7 +13,8 @@ use Nip\Container\Container;
 use Psr\Log\InvalidArgumentException;
 
 /**
- * Class MonologWrappers
+ * Trait MonologWrappers
+ *
  * @package Nip\Logger\Manager
  *
  * @method Container getContainer()
@@ -22,18 +24,17 @@ trait MonologWrappers
     /**
      * Create an instance of any handler available in Monolog.
      *
-     * @param array $config
-     * @return \Psr\Log\LoggerInterface
+     * @param array<string, mixed>|Config $config
      *
      * @throws \InvalidArgumentException
      */
-    protected function createMonologDriver($config)
+    protected function createMonologDriver(array|Config $config): \Psr\Log\LoggerInterface
     {
         $config = $config instanceof Config ? $config->toArray() : $config;
 
         if (!is_a($config['handler'], HandlerInterface::class, true)) {
             throw new InvalidArgumentException(
-                $config['handler'].' must be an instance of '.HandlerInterface::class
+                $config['handler'] . ' must be an instance of ' . HandlerInterface::class
             );
         }
 
@@ -52,12 +53,12 @@ trait MonologWrappers
     }
 
     /**
-     * Prepare the handlers for usage by Monolog.
+     * Prepare multiple handlers for usage by Monolog.
      *
-     * @param array $handlers
-     * @return array
+     * @param HandlerInterface[] $handlers
+     * @return HandlerInterface[]
      */
-    protected function prepareHandlers(array $handlers)
+    protected function prepareHandlers(array $handlers): array
     {
         foreach ($handlers as $key => $handler) {
             $handlers[$key] = $this->prepareHandler($handler);
@@ -67,37 +68,34 @@ trait MonologWrappers
     }
 
     /**
-     * Prepare the handler for usage by Monolog.
+     * Prepare a single handler for usage by Monolog.
      *
-     * @param \Monolog\Handler\HandlerInterface $handler
-     * @param array $config
-     * @return \Monolog\Handler\HandlerInterface
+     * Sets the default LineFormatter on any FormattableHandlerInterface unless
+     * the config provides a custom formatter class name.
+     *
+     * @param array<string, mixed> $config
      */
-    protected function prepareHandler(HandlerInterface $handler, array $config = [])
+    protected function prepareHandler(HandlerInterface $handler, array $config = []): HandlerInterface
     {
-        $isHandlerFormattable = false;
-
-        if (Monolog::API === 1) {
-            $isHandlerFormattable = true;
-        } elseif (Monolog::API === 2 && $handler instanceof FormattableHandlerInterface) {
-            $isHandlerFormattable = true;
+        if (!($handler instanceof FormattableHandlerInterface)) {
+            return $handler;
         }
 
-        if ($isHandlerFormattable && !isset($config['formatter'])) {
+        if (!isset($config['formatter'])) {
             $handler->setFormatter($this->formatter());
-        } elseif ($isHandlerFormattable && $config['formatter'] !== 'default') {
-            $handler->setFormatter($this->app->make($config['formatter'], $config['formatter_with'] ?? []));
+        } elseif ($config['formatter'] !== 'default') {
+            $handler->setFormatter(
+                $this->getContainer()->make($config['formatter'], $config['formatter_with'] ?? [])
+            );
         }
 
         return $handler;
     }
 
     /**
-     * Get a Monolog formatter instance.
-     *
-     * @return \Monolog\Formatter\FormatterInterface
+     * Get a default Monolog LineFormatter instance.
      */
-    protected function formatter()
+    protected function formatter(): \Monolog\Formatter\FormatterInterface
     {
         $formatter = new LineFormatter(null, $this->dateFormat, true, true);
         $formatter->includeStacktraces();
@@ -105,3 +103,4 @@ trait MonologWrappers
         return $formatter;
     }
 }
+